requirements.txt
python-dotenv
django
mysqlclient
djangorestframework
django-cors-headers
django-filter
markdown
djangorestframework-simplejwt
datetime
django-storages[azure]